+++ packages/contracts/src/test/permission/PermissionConditionMock.sol
@@ -5,18 +5,23 @@ pragma solidity 0.8.17;
 import "../../core/permission/PermissionCondition.sol";
 
 contract PermissionConditionMock is PermissionCondition {
-    bool internal _hasPermissionsResult = true;
+    bool public answer;
 
-    function isGranted(
-        address /* _where */,
-        address /* _who */,
-        bytes32 /* _permissionId */,
-        bytes memory /* _data */
-    ) external view returns (bool) {
-        return _hasPermissionsResult;
+    constructor() {
+        answer = true;
     }
 
-    function setWillPerform(bool _result) external {
-        _hasPermissionsResult = _result;
+    function setAnswer(bool _answer) external {
+        answer = _answer;
+    }
+
+    function isGranted(
+        address _where,
+        address _who,
+        bytes32 _permissionId,
+        bytes memory _data
+    ) external view returns (bool) {
+        (_where, _who, _permissionId, _data);
+        return answer;
     }
 }